requirements_conda.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=1_gnu
aom=3.3.0=h27087fc_1
appdirs=1.4.4=pypi_0
argon2-cffi=21.3.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
asttokens=2.0.5=pypi_0
attrs=21.4.0=pypi_0
audioread=2.1.9=pypi_0
backcall=0.2.0=pypi_0
beautifulsoup4=4.10.0=pypi_0
blas=1.0=mkl
bleach=4.1.0=pypi_0
brotli=1.0.9=h7f98852_6
brotli-bin=1.0.9=h7f98852_6
bzip2=1.0.8=h7f98852_4
ca-certificates=2021.10.8=ha878542_0
cachetools=5.0.0=pypi_0
certifi=2021.10.8=pypi_0
cffi=1.15.0=pypi_0
chainer=7.8.1=pypi_0
charset-normalizer=2.0.12=pypi_0
ci-sdr=0.0.0=pypi_0
click=8.0.4=pypi_0
cloudpickle=2.0.0=pyhd8ed1ab_0
configargparse=1.5.3=pypi_0
ctc-segmentation=1.7.1=pypi_0
cudatoolkit=11.3.1=h2bc3f7f_2
cycler=0.11.0=pyhd8ed1ab_0
cython=0.29.28=pypi_0
cytoolz=0.11.2=py38h497a2fe_1
dask-core=2022.2.1=pyhd8ed1ab_0
debugpy=1.5.1=pypi_0
decorator=5.1.1=pypi_0
defusedxml=0.7.1=pypi_0
dill=0.3.4=pypi_0
distance=0.1.3=pypi_0
editdistance=0.6.0=pypi_0
einops=0.4.1=pypi_0
entrypoints=0.4=pypi_0
enum34=1.1.10=py38h32f6830_2
espnet=0.10.6=pypi_0
espnet-tts-frontend=0.0.3=pypi_0
executing=0.8.3=pypi_0
ffmpeg=5.0.0=h594f047_1
filelock=3.6.0=pypi_0
flatbuffers=2.0=pypi_0
fonttools=4.29.1=pypi_0
freetype=2.10.4=h0708190_1
fsspec=2022.2.0=pyhd8ed1ab_0
g2p-en=2.1.0=pypi_0
gdown=4.4.0=pypi_0
giflib=5.2.1=h36c2ea0_2
glob2=0.7=pypi_0
gmp=6.2.1=h58526e2_0
gnutls=3.6.13=h85f3911_1
google-auth=2.6.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
humanfriendly=10.0=pypi_0
icu=69.1=h9c3ff4c_0
idna=3.3=pypi_0
imagecodecs-lite=2019.12.3=py38h6c62de6_4
imageio=2.16.1=pyhcf75d05_0
importlib-metadata=4.11.2=pypi_0
importlib-resources=5.4.0=pypi_0
inflect=5.4.0=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
ipykernel=6.9.1=pypi_0
ipython=8.1.1=pypi_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=7.6.5=pypi_0
jaconv=0.3=pypi_0
jamo=0.4.1=pypi_0
jbig=2.1=h7f98852_2003
jedi=0.18.1=pypi_0
jinja2=3.0.3=pypi_0
joblib=1.1.0=pypi_0
jpeg=9e=h7f98852_0
jsonpatch=1.32=pypi_0
jsonpointer=2.2=pypi_0
jsonschema=4.4.0=pypi_0
jupyter=1.0.0=pypi_0
jupyter-client=7.1.2=pypi_0
jupyter-console=6.4.3=pypi_0
jupyter-core=4.9.2=pypi_0
jupyterlab-pygments=0.1.2=pypi_0
jupyterlab-widgets=1.0.2=pypi_0
kaldiio=2.17.2=pypi_0
keras=2.8.0=pypi_0
kiwisolver=1.3.2=py38h1fd1430_1
lame=3.100=h7f98852_1001
lcms2=2.12=hddcbb42_0
ld_impl_linux-64=2.36.1=hea4e1c9_2
lerc=3.0=h9c3ff4c_0
libbrotlicommon=1.0.9=h7f98852_6
libbrotlidec=1.0.9=h7f98852_6
libbrotlienc=1.0.9=h7f98852_6
libclang=13.0.0=pypi_0
libdeflate=1.10=h7f98852_0
libdrm=2.4.109=h7f98852_0
libffi=3.4.2=h7f98852_5
libgcc-ng=11.2.0=h1d223b6_13
libgfortran-ng=7.5.0=h14aa051_20
libgfortran4=7.5.0=h14aa051_20
libgomp=11.2.0=h1d223b6_13
libiconv=1.16=h516909a_0
libnsl=2.0.0=h7f98852_0
libpciaccess=0.16=h516909a_0
libpng=1.6.37=h21135ba_2
librosa=0.9.1=pypi_0
libstdcxx-ng=11.2.0=he4da1e4_13
libtiff=4.3.0=h542a066_3
libuv=1.43.0=h7f98852_0
libva=2.14.0=h7f98852_0
libvpx=1.11.0=h9c3ff4c_3
libwebp=1.2.2=h3452ae3_0
libwebp-base=1.2.2=h7f98852_1
libxcb=1.13=h7f98852_1004
libxml2=2.9.12=h885dcf4_1
libzlib=1.2.11=h36c2ea0_1013
llvmlite=0.38.0=pypi_0
locket=0.2.0=py_2
lz4-c=1.9.3=h9c3ff4c_1
markdown=3.3.6=pypi_0
markupsafe=2.1.0=pypi_0
matplotlib-base=3.5.1=py38hf4fb855_0
matplotlib-inline=0.1.3=pypi_0
mistune=0.8.4=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py38h95df7f1_0
mkl_fft=1.3.1=py38h8666266_1
mkl_random=1.2.2=py38h1abd341_0
multiprocess=0.70.12.2=pypi_0
munkres=1.1.4=pyh9f0ad1d_0
nbclient=0.5.12=pypi_0
nbconvert=6.4.2=pypi_0
nbformat=5.1.3=pypi_0
ncurses=6.3=h9c3ff4c_0
nest-asyncio=1.5.4=pypi_0
nettle=3.6=he412f7d_0
networkx=2.7.1=pyhd8ed1ab_0
nltk=3.7=pypi_0
notebook=6.4.8=pypi_0
numba=0.55.1=pypi_0
numpy=1.21.5=pypi_0
numpy-base=1.21.2=py38h79a1101_0
oauthlib=3.2.0=pypi_0
openh264=2.1.1=h780b84a_0
openjpeg=2.4.0=hb52868f_1
openssl=3.0.0=h7f98852_2
packaging=21.3=pyhd8ed1ab_0
pandas=1.4.1=py38h43a58ef_0
pandocfilters=1.5.0=pypi_0
parso=0.8.3=pypi_0
partd=1.2.0=pyhd8ed1ab_0
pathlib=1.0.1=py38h578d9bd_5
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=9.0.1=py38h0ee0e06_2
pip=22.0.4=pyhd8ed1ab_0
pooch=1.6.0=pypi_0
prometheus-client=0.13.1=pypi_0
prompt-toolkit=3.0.28=pypi_0
pthread-stubs=0.4=h36c2ea0_1001
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.21=pypi_0
pygments=2.11.2=pypi_0
pykaldi=0.2.0=pypi_0
pynndescent=0.5.6=pypi_0
pyparsing=3.0.7=pyhd8ed1ab_0
pypinyin=0.44.0=pypi_0
pyqt5=5.15.6=pypi_0
pyqt5-qt5=5.15.2=pypi_0
pyqt5-sip=12.9.1=pypi_0
pyrsistent=0.18.1=pypi_0
pysocks=1.7.1=pypi_0
python=3.8.12=h0744224_3_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.8=2_cp38
pytorch=1.10.2=py3.8_cuda11.3_cudnn8.2.0_0
pytorch-mutex=1.0=cuda
pytorch-ranger=0.1.1=pypi_0
pytorch-wpe=0.0.1=pypi_0
pytz=2021.3=pyhd8ed1ab_0
pywavelets=1.3.0=py38h3ec907f_0
pyworld=0.3.0=pypi_0
pyyaml=6.0=py38h497a2fe_3
pyzmq=22.3.0=pypi_0
qtconsole=5.2.2=pypi_0
qtpy=2.0.1=pypi_0
readline=8.1=h46c0cb4_0
regex=2022.3.2=pypi_0
requests=2.27.1=pypi_0
requests-oauthlib=1.3.1=pypi_0
resampy=0.2.2=pypi_0
rsa=4.8=pypi_0
scikit-image=0.19.2=py38h43a58ef_0
scikit-learn=1.0.2=pypi_0
scipy=1.8.0=pypi_0
send2trash=1.8.0=pypi_0
sentencepiece=0.1.96=pypi_0
setuptools=59.5.0=pypi_0
six=1.16.0=pyh6c4a22f_0
sounddevice=0.4.4=pypi_0
soundfile=0.10.3.post1=pypi_0
soupsieve=2.3.1=pypi_0
sqlite=3.37.0=h9cd32fc_0
stack-data=0.2.0=pypi_0
svt-av1=0.9.1=h27087fc_0
tensorboard=2.8.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
tensorflow=2.8.0=pypi_0
tensorflow-addons=0.12.1=pypi_0
tensorflow-io-gcs-filesystem=0.24.0=pypi_0
terminado=0.13.3=pypi_0
testpath=0.6.0=pypi_0
tf-estimator-nightly=2.8.0.dev2021122109=pypi_0
threadpoolctl=3.1.0=pypi_0
tifffile=2019.7.26.2=py38_0
tk=8.6.12=h27826a3_0
toolz=0.11.2=pyhd8ed1ab_0
torch-complex=0.4.3=pypi_0
torch-optimizer=0.3.0=pypi_0
torchaudio=0.10.2=py38_cu113
torchfile=0.1.0=pypi_0
torchvision=0.11.3=py38_cu113
tornado=6.1=pypi_0
tqdm=4.63.0=pypi_0
traitlets=5.1.1=pypi_0
typeguard=2.13.3=pypi_0
typing_extensions=4.1.1=pyha770c72_0
umap-learn=0.5.2=pypi_0
unicodedata2=14.0.0=py38h497a2fe_0
unidecode=1.3.3=pypi_0
urllib3=1.26.8=pypi_0
visdom=0.1.8.9=pypi_0
wcwidth=0.2.5=pypi_0
webencodings=0.5.1=pypi_0
webrtcvad=2.0.10=pypi_0
websocket-client=1.3.1=pypi_0
werkzeug=2.0.3=pypi_0
wheel=0.37.1=pyhd8ed1ab_0
widgetsnbextension=3.5.2=pypi_0
x264=1!161.3030=h7f98852_1
x265=3.5=h924138e_3
xorg-fixesproto=5.0=h7f98852_1002
xorg-kbproto=1.0.7=h7f98852_1002
xorg-libx11=1.7.2=h7f98852_0
xorg-libxau=1.0.9=h7f98852_0
xorg-libxdmcp=1.1.3=h7f98852_0
xorg-libxext=1.3.4=h7f98852_1
xorg-libxfixes=5.0.3=h7f98852_1004
xorg-xextproto=7.3.0=h7f98852_1002
xorg-xproto=7.0.31=h7f98852_1007
xz=5.2.5=h516909a_1
yaml=0.2.5=h7f98852_2
zipp=3.7.0=pypi_0
zlib=1.2.11=h36c2ea0_1013
zstd=1.5.2=ha95c52a_0
